Improve testing TCs that are automated only with 'su'
First automated tests used 'su' as a mean of testing smart card authentication. However, these tests were usually supposed to test login (not su), but support for any direct or indirect testing of login was not implemented yet. Testing logins in TTY and also in GUI is already supported. It's time to check test coverage and cover TCs that still use 'su' also with TTY and GUI logins.